tests/requirements.txt
boto3
pytest
python-dotenv
requests
